Changed order of positional arguments to <left> <base> <right>
This matches the synopsis of merge(1) and (more importantly) is also
what mercurial assumes when calling a merge-tool for which <tool>.args
hasn't been set. So this change should allow to hook into hg with less
configuration work required from the user.

The previous argument order matched that of KDiff3 (but not diff3(1),
which also uses the same order as merge(1)), but there is no technical
reason to follow that.

Contribmerge will probably most often be called automatically from
either hg or git, so it makes sense to follow their defaults. git
doesn't seem to have one (requires explicit configuration of the
complete command line for each used mergetool, if not already known), so
we use hg's.