test merging (WIP): put contribmerge binary into MERGE_TEST_DRIVER variable
The 3rd argument to ADD_TEST (first argument to the test script passed
as 2nd argument) will be the same for all tests that have
"${MERGE_TEST_DRIVER}" as 2nd argument, namely it will be the full path
to the contribmerge executable. Thus, we can just make it part of the
MERGE_TEST_DRIVER variable.

But it is important that it's a separate string in that variable, or
CTest would think the whole thing is the path to the test driver script.